+++ packages/learn-card-core/src/wallet/plugins/vc/issuePresentation.ts
@@ -1,26 +1,25 @@
 import { UnsignedVP } from '@learncard/types';
 
 import { ProofOptions } from '@wallet/plugins/didkit/types';
-import { VCImplicitWallet, VCPluginDependentMethods } from './types';
-import { Wallet } from 'types/wallet';
+import { VCDependentLearnCard, VCImplicitLearnCard } from './types';
 
-export const issuePresentation = (initWallet: Wallet<string, VCPluginDependentMethods>) => {
+export const issuePresentation = (initLearnCard: VCDependentLearnCard) => {
     return async (
-        wallet: VCImplicitWallet,
+        learnCard: VCImplicitLearnCard,
         presentation: UnsignedVP,
         signingOptions: Partial<ProofOptions> = {}
     ) => {
-        const kp = wallet.pluginMethods.getSubjectKeypair();
+        const kp = learnCard.id.keypair();
 
         if (!kp) throw new Error('Cannot issue credential: Could not get subject keypair');
 
         const options = {
-            verificationMethod: await initWallet.pluginMethods.keyToVerificationMethod('key', kp),
+            verificationMethod: await initLearnCard.invoke.keyToVerificationMethod('key', kp),
             proofPurpose: 'assertionMethod',
             type: 'Ed25519Signature2020',
             ...signingOptions,
         };
 
-        return initWallet.pluginMethods.issuePresentation(presentation, options, kp);
+        return initLearnCard.invoke.issuePresentation(presentation, options, kp);
     };
 };