+++ packages/learn-card-types/src/index.ts
@@ -1,5 +1,4 @@
 export * from './vc';
 export * from './obv3';
 export * from './learncard';
-export * from './idx';
 export * from './crypto';