+++ examples/chapi-example/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @learncard/chapi-example
 
+## 1.0.19
+
+### Patch Changes
+
+-   Updated dependencies [[`e7cfb63`](https://github.com/learningeconomy/LearnCard/commit/e7cfb636b21cfdd834e3b0cb028036819326a2f9), [`e7cfb63`](https://github.com/learningeconomy/LearnCard/commit/e7cfb636b21cfdd834e3b0cb028036819326a2f9)]:
+    -   @learncard/core@8.0.0
+    -   @learncard/react@2.3.49
+
+## 1.0.18
+
+### Patch Changes
+
+-   Updated dependencies [[`b655513`](https://github.com/learningeconomy/LearnCard/commit/b6555139a5fcf653db514d8ee059bd9edd99fd54)]:
+    -   @learncard/react@2.3.48
+
 ## 1.0.17
 
 ### Patch Changes