+++ examples/snap-example-dapp/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.0.12
+
+### Patch Changes
+
+-   Updated dependencies [[`e7cfb63`](https://github.com/learningeconomy/LearnCard/commit/e7cfb636b21cfdd834e3b0cb028036819326a2f9), [`e7cfb63`](https://github.com/learningeconomy/LearnCard/commit/e7cfb636b21cfdd834e3b0cb028036819326a2f9)]:
+    -   @learncard/core@8.0.0
+    -   @learncard/meta-mask-snap@1.0.12
+
 ## 1.0.11
 
 ### Patch Changes